+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `specs`
 
 ### Changed
+* `IntoIterator` compatibility for args in `execute()`
 * updated: runnel(0.4.0)
 * updated: rust-version-info-file(0.2.0)
 * updated: regex(1.11)